[Test] Further bounds testing (#214)
More detailed tests of the bounds on `integer` and `number` types. As part of this, the `json_schema_check()` routine has been reworked to match the requirements of the JSON testing case (basically, each incoming test `Value` is complete, so if the grammar hasn't matched by the end, there's a problem).

The `integer` bounds tests all work. Unfortunately, as detailed in #216 , there do appear to be residual problems in the `number` bounds. For this reason, I have not expanded the more comprehensive test approach with a `number_both_bounds()` test; there is little point until the individual bounds tests work in full.